Refactor: company command and query tests
